connecting database MongoDb

+++ src/controllers/personagem.controller.js
@@ -1,7 +1,7 @@
 const personagensService = require('../services/personagem.service');
 
 const findAllPersonagensController = async (req, res) => {
-  const personagens = await personagensService.findAllPersonagensService();
+  const personagens = personagensService.findAllPersonagensService();
   if (personagens.length == 0) {
     return res.status(404).send({ message: 'Lista de personagens vazia' });
   }
@@ -9,9 +9,12 @@ const findAllPersonagensController = async (req, res) => {
 };
 const findByIdPersonagemController = async (req, res) => {
   const parametroId = Number(req.params.id);
-  const choosePersonagem = await personagensService.findByIdPersonagemService(
-    parametroId,
-  );
+
+  if (!parametroId) {
+    return res.status(400).send({ message: 'id inválido' });
+  }
+  const choosePersonagem =
+    personagensService.findByIdPersonagemService(parametroId);
   if (!choosePersonagem) {
     return res.status(404).send({ message: 'Personagem não encontrada' });
   }
@@ -20,13 +23,45 @@ const findByIdPersonagemController = async (req, res) => {
 
 const createPersonagemController = (req, res) => {
   const personagem = req.body;
+  if (
+    !personagem ||
+    !personagem.nome ||
+    !personagem.personagem ||
+    !personagem.descricao ||
+    !personagem.imagem ||
+    !personagem.idade ||
+    !personagem.nacionalidade ||
+    !personagem.carreira
+  ) {
+    return res
+      .status(400)
+      .send({ message: 'Envie todos os campos do personagem' });
+  }
   const newPersonagem = personagensService.createPersonagemService(personagem);
-  res.send(newPersonagem);
+  res.status(201).send(newPersonagem);
 };
 
 const updatePersonagemController = (req, res) => {
   const idParam = Number(req.params.id);
+  if (!idParam) {
+    return res.status(400).send({ message: 'id inválido' });
+  }
   const personagemEdit = req.body;
+
+  if (
+    !personagemEdit ||
+    !personagemEdit.nome ||
+    !personagemEdit.personagem ||
+    !personagemEdit.descricao ||
+    !personagemEdit.imagem ||
+    !personagemEdit.idade ||
+    !personagemEdit.nacionalidade ||
+    !personagemEdit.carreira
+  ) {
+    return res
+      .status(400)
+      .send({ message: 'Envie todos os campos do personagem' });
+  }
   const updatedPersonagem = personagensService.updatePersonagemService(
     idParam,
     personagemEdit,
@@ -35,7 +70,11 @@ const updatePersonagemController = (req, res) => {
 };
 
 const deletePersonagemController = (req, res) => {
-  const idParam = req.params.id;
+  const idParam = Number(req.params.id);
+  if (!idParam) {
+    return res.status(400).send({ message: 'id inválido' });
+  }
+
   personagensService.deletePersonagemService(idParam);
   res.send({ message: 'Personagem deletada com sucesso!' });
 };
